requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
bcrypt==4.2.1
    # via paramiko
certifi==2024.7.4
    # via requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
cryptography==44.0.1
    # via paramiko
fabric==2.7.1
    # via
    #   -r requirements.in
    #   patchwork
idna==3.7
    # via requests
invoke==1.7.3
    # via fabric
jq==1.7.0
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
marshmallow==3.26.1
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
packaging==24.2
    # via marshmallow
paramiko==3.5.1
    # via fabric
patchwork==1.0.1
    # via -r requirements.in
pathlib2==2.3.7.post1
    # via fabric
pycparser==2.22
    # via cffi
pydash==8.0.1
    # via -r requirements.in
pygments==2.19.1
    # via rich
pynacl==1.5.0
    # via paramiko
python-dateutil==2.9.0.post0
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
rich==13.9.4
    # via -r requirements.in
six==1.17.0
    # via
    #   pathlib2
    #   python-dateutil
timeago==1.0.16
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   pydash
    #   rich
urllib3==2.2.2
    # via requests
